+++ src/components/instruments/Metronome.tsx
