+++ src/utils/MIDI.ts
@@ -1,0 +1,72 @@
+const MIDI = [
+	{ value: 12, text: "C0" },
+	{ value: 14, text: "D0" },
+	{ value: 16, text: "E0" },
+	{ value: 17, text: "F0" },
+	{ value: 19, text: "G0" },
+	{ value: 21, text: "A0" },
+	{ value: 23, text: "B0" },
+	{ value: 24, text: "C1" },
+	{ value: 26, text: "D1" },
+	{ value: 28, text: "E1" },
+	{ value: 29, text: "F1" },
+	{ value: 31, text: "G1" },
+	{ value: 33, text: "A1" },
+	{ value: 35, text: "B1" },
+	{ value: 36, text: "C2" },
+	{ value: 38, text: "D2" },
+	{ value: 40, text: "E2" },
+	{ value: 41, text: "F2" },
+	{ value: 43, text: "G2" },
+	{ value: 45, text: "A2" },
+	{ value: 47, text: "B2" },
+	{ value: 48, text: "C3" },
+	{ value: 50, text: "D3" },
+	{ value: 52, text: "E3" },
+	{ value: 53, text: "F3" },
+	{ value: 55, text: "G3" },
+	{ value: 57, text: "A3" },
+	{ value: 59, text: "B3" },
+	{ value: 60, text: "C4" },
+	{ value: 62, text: "D4" },
+	{ value: 64, text: "E4" },
+	{ value: 65, text: "F4" },
+	{ value: 67, text: "G4" },
+	{ value: 69, text: "A4" },
+	{ value: 71, text: "B4" },
+	{ value: 72, text: "C5" },
+	{ value: 74, text: "D5" },
+	{ value: 76, text: "E5" },
+	{ value: 77, text: "F5" },
+	{ value: 79, text: "G5" },
+	{ value: 81, text: "A5" },
+	{ value: 83, text: "B5" },
+	{ value: 84, text: "C6" },
+	{ value: 86, text: "D6" },
+	{ value: 88, text: "E6" },
+	{ value: 89, text: "F6" },
+	{ value: 91, text: "G6" },
+	{ value: 93, text: "A6" },
+	{ value: 95, text: "B6" },
+	{ value: 96, text: "C7" },
+	{ value: 98, text: "D7" },
+	{ value: 100, text: "E7" },
+	{ value: 101, text: "F7" },
+	{ value: 103, text: "G7" },
+	{ value: 105, text: "A7" },
+	{ value: 107, text: "B7" },
+	{ value: 108, text: "C8" },
+	{ value: 110, text: "D8" },
+	{ value: 112, text: "E8" },
+	{ value: 113, text: "F8" },
+	{ value: 115, text: "G8" },
+	{ value: 117, text: "A8" },
+	{ value: 119, text: "B8" },
+	{ value: 120, text: "C9" },
+	{ value: 122, text: "D9" },
+	{ value: 124, text: "E9" },
+	{ value: 125, text: "F9" },
+	{ value: 127, text: "G9" },
+];
+
+export default MIDI;
