+++ src/utils/instruments.ts
@@ -1,35 +1,42 @@
 "use client";
+
 import { Soundfont, DrumMachine } from "smplr";
 import {
 	MidiNumbers,
 	//@ts-ignore
 } from "react-piano";
 
-const context = new AudioContext();
+const context = typeof window !== "undefined" ? new AudioContext() : null;
 
-export const piano = new Soundfont(context, {
-	instrument: "acoustic_grand_piano",
-});
+export const piano = context
+	? new Soundfont(context!, {
+			instrument: "acoustic_grand_piano",
+	  })
+	: null;
 
-export const drum = new DrumMachine(context, { instrument: "TR-808" });
+export const drum = context
+	? new DrumMachine(context!, { instrument: "TR-808" })
+	: null;
 
 export const playPiano = (
 	music: Array<{ note: number; start: number; end: number }>,
 	bpm: number
 ) => {
-	const time = 7500 / bpm;
-	for (let i = 0; i < music.length; i++) {
-		const note = music[i].note;
-		const start = music[i].start;
-		const end = music[i].end;
-		setTimeout(() => {
-			piano.start({
-				note: MidiNumbers.getAttributes(note).note,
-			});
-		}, start * time);
-		setTimeout(() => {
-			piano.stop(note);
-		}, end * time);
+	if (context && piano) {
+		const time = 7500 / bpm;
+		for (let i = 0; i < music.length; i++) {
+			const note = music[i].note;
+			const start = music[i].start;
+			const end = music[i].end;
+			setTimeout(() => {
+				piano.start({
+					note: MidiNumbers.getAttributes(note).note,
+				});
+			}, start * time);
+			setTimeout(() => {
+				piano.stop(note);
+			}, end * time);
+		}
 	}
 };
 
@@ -37,16 +44,18 @@ export const playDrum = (
 	music: Array<{ note: string; start: number }>,
 	bpm: number
 ) => {
-	const now = context.currentTime;
-	console.log(music);
-	const time = 30 / bpm;
-	for (let i = 0; i < music.length; i++) {
-		const note = music[i].note;
-		const start = music[i].start;
-		console.log(`playing ${music[i].note}`);
-		drum.start({
-			note: note,
-			time: now + start * time,
-		});
+	if (context && drum) {
+		const now = context!.currentTime;
+		console.log(music);
+		const time = 30 / bpm;
+		for (let i = 0; i < music.length; i++) {
+			const note = music[i].note;
+			const start = music[i].start;
+			console.log(`playing ${music[i].note}`);
+			drum.start({
+				note: note,
+				time: now + start * time,
+			});
+		}
 	}
 };
